+++ src/service/export/abstractExportService.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import { ExportService } from "./exportService";
-import { FieldInfo } from "mysql";
+import { FieldInfo } from "mysql2";
 import * as fs from "fs";
 import { Console } from "../../common/Console";
 import { ExportOption } from "./exportOption";