+++ src/view/status/mysqlStatusService.ts
@@ -1,4 +1,4 @@
-import { FieldInfo } from "mysql";
+import { FieldInfo } from "mysql2";
 import { Console } from "../../common/Console";
 import { ConnectionNode } from "../../model/database/connectionNode";
 import { ConnectionManager } from "../../service/connectionManager";